feat: change the positions of the function spa_deriv.
